refactor: include polyfill into flatten util
